Migrated to distinct TodoApp component
